requirements.txt
torch>=1.10.0
ffmpeg-python>=0.2.0
faster-whisper>=0.9.0
pyannote.audio>=0.0.1
transformers>=4.30.0
numpy>=1.24.0
python-dotenv>=0.9.9
spacy>=3.5.0
sarvamai
langchain-google-genai
streamlit
python-docx